Concluding the portfolio fragment for now. All works i just may have to look at the mechanics of if to ensure the use case is viable or needs to be changed to show daily changes only? Changed Doubles to BigDecimals to try and solve an issue with the changes of the adapter in portfoliofragment not matching up with the combined change at the top. This was actually due to me multiplying the change by the chosenfiat and therfore giving it more decimal points to which it obtained a more accurate answer compared to the adapter items. Took bloody ages lmao
You'll have to read commit changes for the rest (dont think theres much else) as its been a long day and i cant remember lmao

Peace xo

+++ app/src/main/java/com/jonnycaley/cryptomanager/data/model/CryptoCompare/HistoricalData/Datum.kt
@@ -3,6 +3,7 @@ package com.jonnycaley.cryptomanager.data.model.CryptoCompare.HistoricalData
 import java.io.Serializable
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 class Datum : Serializable {
 
@@ -11,16 +12,16 @@ class Datum : Serializable {
     var time: Long? = null
     @SerializedName("close")
     @Expose
-    var close: Double? = null
+    var close: BigDecimal? = null
     @SerializedName("high")
     @Expose
-    var high: Double? = null
+    var high: BigDecimal? = null
     @SerializedName("low")
     @Expose
-    var low: Double? = null
+    var low: BigDecimal? = null
     @SerializedName("open")
     @Expose
-    var open: Double? = null
+    var open: BigDecimal? = null
     @SerializedName("volumefrom")
     @Expose
     var volumefrom: Double? = null
