+++ app/src/main/java/com/jonnycaley/cryptomanager/data/model/CryptoCompare/MultiPrice/Prices.kt
@@ -3,12 +3,13 @@ package com.jonnycaley.cryptomanager.data.model.CryptoCompare.MultiPrice
 import java.io.Serializable
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 class Prices : Serializable {
 
     @SerializedName("USD")
     @Expose
-    var uSD: Double? = null
+    var uSD: BigDecimal? = null
 //    @SerializedName("BTC")
 //    @Expose
 //    var bTC: Double? = null
