+++ app/src/main/java/com/jonnycaley/cryptomanager/data/model/ExchangeRates/Rate.kt
@@ -3,6 +3,7 @@ package com.jonnycaley.cryptomanager.data.model.ExchangeRates
 import java.io.Serializable
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 class Rate : Serializable {
 
@@ -11,7 +12,7 @@ class Rate : Serializable {
     var fiat: String? = null
     @SerializedName("rate")
     @Expose
-    var rate: Double? = null
+    var rate: BigDecimal? = null
 
     companion object {
         private const val serialVersionUID = 1944469845247308751L
